requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.1.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.12.3
    # via -r requirements.in
bottle==0.13.2
    # via -r requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
cffi==1.17.1
    # via trio
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
jiter==0.8.2
    # via openai
numpy==2.2.2
    # via pandas
openai==1.60.2
    # via -r requirements.in
outcome==1.3.0.post0
    # via trio
packaging==24.2
    # via webdriver-manager
pandas==2.2.3
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.0.1
    # via webdriver-manager
pytz==2024.2
    # via pandas
requests==2.32.3
    # via
    #   -r requirements.in
    #   webdriver-manager
selenium==4.28.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.6
    # via beautifulsoup4
tqdm==4.67.1
    # via openai
trio==0.28.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   selenium
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r requirements.in
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
